backend/rozetka_api/requirements.txt
asgiref==3.6.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1
djangorestframework==3.12.4
env==0.1.0
environs==9.5.0
filelock==3.12.0
marshmallow==3.19.0
packaging==23.1
Pillow==9.4.0
platformdirs==3.5.1
psycopg2==2.9.6
psycopg2-binary==2.9.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
virtualenv==20.23.0
djangorestframework-simplejwt==5.2.2
